feat: BeerSearchList component added and used for review
