+++ src/components/shared/ReviewModal/BeerSearchContent.tsx
@@ -1,15 +1,12 @@
+import {SearchBarList} from "@/components/home/Search/SearchBarList";
 import {
-  Box,
   ModalBody,
   ModalContent,
   ModalContentProps,
   ModalHeader,
-  Text,
   VStack,
 } from "@chakra-ui/react";
-import React, {useState} from "react";
-import {BeerResultType, CategoryType} from "../../../../interface/types";
-import SearchInput from "../../search/SearchInput";
+import React from "react";
 import {LeftBackRandom} from "../Headers/LeftBackRandom";
 
 interface BeerSearchContentProps extends ModalContentProps {
@@ -22,41 +19,13 @@ export const BeerSearchContent: React.FC<BeerSearchContentProps> = ({
   onChangeBeerName,
   ...props
 }) => {
-  const [value, setValue] = useState<string>("");
-  // recoil state가 되어야 함.
-  const [allBeers, setAllBeers] = useState<BeerResultType[]>([]);
-  const MOCK_CATEGORY = [
-    {
-      id: 0,
-      name_ko: "에일",
-      name_en: "ale",
-      description: "This is ale",
-    },
-    {
-      id: 1,
-      name_ko: "라거",
-      name_en: "Lagar",
-      description: "This is Lagar",
-    },
-  ];
-  const [selectedItems, setSelectedItems] = useState<any[]>(MOCK_CATEGORY);
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setValue(e.target.value);
-    // setSelectedItems(filterSearchResult(e.target.value));
-  };
-
-  const clearValue = () => {
-    setValue("");
-  };
-
-  const handleClick = (beerName: string) => {
+  const handleClick = (name: string) => {
     onClickBack();
-    onChangeBeerName(beerName);
+    onChangeBeerName(name);
   };
 
   return (
-    <ModalContent px="20px" {...props}>
+    <ModalContent px="20px" {...props} bg="white">
       <ModalHeader pt="46px">
         <LeftBackRandom onClick={onClickBack} title="맥주 이름" />
       </ModalHeader>
@@ -67,25 +36,7 @@ export const BeerSearchContent: React.FC<BeerSearchContentProps> = ({
           justifyContent="flex-start"
           alignItems={"flex-start"}
         >
-          {/* search bar  */}
-          <SearchInput onChange={handleChange} clearValue={clearValue} />
-          {/* search results */}
-          <VStack px="10px">
-            {selectedItems.map((item: CategoryType) => {
-              return (
-                <Box
-                  key={item.id}
-                  py="10px"
-                  onClick={() => handleClick(item.name_ko)}
-                  cursor="pointer"
-                >
-                  <Text textStyle="h2" textColor={"black.100"}>
-                    {item.name_ko}
-                  </Text>
-                </Box>
-              );
-            })}
-          </VStack>
+          <SearchBarList handleClickItem={handleClick} />
         </VStack>
       </ModalBody>
     </ModalContent>
